+++ src/IRremote.h
@@ -21,7 +21,7 @@
 //------------------------------------------------------------------------------
 // The ISR header contains several useful macros the user may wish to use
 //
-#include "IRremoteInt.h"
+#include "private/IRremoteInt.h"
 
 //------------------------------------------------------------------------------
 // Supported IR protocols
