+++ src/irRecv.cpp
@@ -1,5 +1,4 @@
 #include "IRremote.h"
-#include "IRremoteInt.h"
 
 //+=============================================================================
 // Decodes the received IR message
