+++ src/ir_Denon.cpp
@@ -1,5 +1,4 @@
 #include "IRremote.h"
-#include "IRremoteInt.h"
 
 // Reverse Engineered by looking at RAW dumps generated by IRremote
