+++ src/ir_Dish.cpp
@@ -1,5 +1,4 @@
 #include "IRremote.h"
-#include "IRremoteInt.h"
 
 //==============================================================================
 //                       DDDD   IIIII   SSSS  H   H
