+++ src/ir_JVC.cpp
@@ -1,5 +1,4 @@
 #include "IRremote.h"
-#include "IRremoteInt.h"
 
 //==============================================================================
 //                             JJJJJ  V   V   CCCC
