+++ src/ir_NEC.cpp
@@ -1,5 +1,4 @@
 #include "IRremote.h"
-#include "IRremoteInt.h"
 
 //==============================================================================
 //                           N   N  EEEEE   CCCC
