+++ src/ir_Template.cpp
@@ -92,7 +92,6 @@ Regards,
 */
 
 #include "IRremote.h"
-#include "IRremoteInt.h"
 
 //==============================================================================
 //
